Add files via upload
This dataset is more focused on EDA in a classical dataset like Automobile Dataset and it will face many aspects about EDA and Data Visualization in the first part. In the second we will try to use Machine Learning algorithm to solve problem in both cases ( Classification and Regression) for the price. We used the price variable as target. For the Classification task we choose a threshold to split the cars in 2 category of price. For the Regression task we just converted the original variable in numeric